Add support funky built-in intents
Defining intent handlers by defining methods whose names
are based on the intent names works fine for simple intent
names, but the Amazon built-in intent names may include
periods, @, <, >, [, and ] - all of which can't appear in
a method name.  So, the solution is to introduce name mangling.

This was already done to an extent, in that the handler method
lowered the case of the intent name before searching for a
matching method name.  This adds slightly more complicated rules
that normalize the built-in intent names.  For names without
any of these complicating features, the mangled name remains
just the lowercase version of the intent name.  You can use
the mangle() method to determine the mangled named of any
intent and use that to name your function - or you can define
these handlers after the class definition using the @intent()
decorator provided by the Skill class.

Assuming you had a custom skill defined as the class MySkill,
then you can add a new intent handler (outside the class
definition) as follows:

@MySkill.intent('AMAZON.HelpIntent')
def my_help_handler(self, request):
    return Response.respond('Some helpful text!')

While this is primarily intented to simplify the addition
of handlers for intents with complicated mangled names,
it can be used to add handler functions for any intent.